Add support for storing which time base a USF file uses
Store which time base a USF file uses in its header. Old files will
default to 'accesses', which is the typical case. Future applications
may want to store cycles or instructions instead of accesses, enable
that by reserving two bits in the header to represent the time base.
Also, update the usfdump utility to display which time base is used.